[Fleet]: add log statement when deleting transform (#81927)
[Fleet]: add log statement when deleting transform
# Conflicts:
#	x-pack/plugins/ingest_manager/server/services/epm/elasticsearch/transform/remove.ts

+++ x-pack/plugins/ingest_manager/server/services/epm/elasticsearch/transform/remove.ts
@@ -23,6 +23,8 @@ export const deleteTransforms = async (
   callCluster: CallESAsCurrentUser,
   transformIds: string[]
 ) => {
+  const logger = appContextService.getLogger();
+  logger.info(`Deleting currently installed transform ids ${transformIds}`);
   await Promise.all(
     transformIds.map(async (transformId) => {
       // get the index the transform
@@ -45,14 +47,18 @@ export const deleteTransforms = async (
         path: `/_transform/${transformId}`,
         ignore: [404],
       });
-
-      // expect this to be 1
-      for (const transform of transformResponse.transforms) {
-        await callCluster('transport.request', {
-          method: 'DELETE',
-          path: `/${transform?.dest?.index}`,
-          ignore: [404],
-        });
+      logger.info(`Deleted: ${transformId}`);
+      if (transformResponse?.transforms) {
+        // expect this to be 1
+        for (const transform of transformResponse.transforms) {
+          await callCluster('transport.request', {
+            method: 'DELETE',
+            path: `/${transform?.dest?.index}`,
+            ignore: [404],
+          });
+        }
+      } else {
+        logger.warn(`cannot find transform for ${transformId}`);
       }
     })
   );
